+++ packages/babel-plugin-fbt/babel-processors/FbtFunctionCallProcessor.js
@@ -16,6 +16,9 @@ import type {
   ValidPronounUsagesKey,
 } from '../FbtConstants';
 import type {
+  ParamSet,
+} from '../FbtUtil';
+import type {
   FbtBabelNodeCallExpression,
   FbtBabelNodeJSXElement,
   FbtBabelNodeShape,
@@ -29,13 +32,23 @@ type NodePath = NodePathOf<FbtBabelNodeCallExpression>;
 export type ExtractTableTextItems = Array<
   | ?string
   | boolean
-  | BabelNodeExpression | BabelNodeSpreadElement | BabelNodeJSXNamespacedName
-  | {| type: 'subject' |}
+  | BabelNodeExpression
+  | BabelNodeSpreadElement
+  | BabelNodeJSXNamespacedName
+  | {|type: 'subject'|}
   | {|
-    type: 'enum',
-    range: {[name: string]: BabelNodeStringLiteral},
-    value: string,
-  |}
+      type: 'number',
+      token: string,
+    |}
+  | {|
+      type: 'gender',
+      token: string,
+    |}
+  | {|
+      type: 'enum',
+      range: {[name: string]: BabelNodeStringLiteral},
+      value: string,
+    |}
   | {|
       type: 'plural',
       count: ?FbtOptionValue,
@@ -46,12 +59,12 @@ export type ExtractTableTextItems = Array<
       many: string,
     |}
   | {|
-    type: 'pronoun',
-    capitalize: boolean,
-    gender: string,
-    human: boolean,
-    usage: ValidPronounUsagesKey,
-  |}
+      type: 'pronoun',
+      capitalize: boolean,
+      gender: string,
+      human: boolean,
+      usage: ValidPronounUsagesKey,
+    |},
 >;
 
 export type FbtFunctionCallPhrase = {|
@@ -66,6 +79,7 @@ const {
   FbtBooleanOptions,
   FbtType,
   PLURAL_PARAM_TOKEN,
+  SENTINEL,
   ValidFbtOptions,
   ValidPluralOptions,
   ValidPronounOptions,
@@ -145,7 +159,7 @@ class FbtFunctionCallProcessor {
     path: NodePath,
     pluginOptions: PluginOptions,
   } */) /*: ?FbtFunctionCallProcessor */ {
-    const nodeChecker = FbtNodeChecker.forFbtFunctionCall(path);
+    const nodeChecker = FbtNodeChecker.forFbtFunctionCall(path.node);
     return nodeChecker != null
       ? new FbtFunctionCallProcessor({
         babelTypes,
@@ -160,7 +174,7 @@ class FbtFunctionCallProcessor {
 
   _assertJSModuleWasAlreadyRequired() /*: this */ {
     const {moduleName, path} = this;
-    if (!this.nodeChecker.isJSModuleBound(path)) {
+    if (!this.nodeChecker.isJSModuleBound<typeof path.node>(path)) {
       throw errorAt(
         path.node,
         `${moduleName} is not bound. Did you forget to require('${moduleName}')?`,
@@ -216,11 +230,20 @@ class FbtFunctionCallProcessor {
   // Returns params and enums info in the order in which they appear.
   _collectFbtCalls(options /*: $Shape<FbtCallSiteOptions> */) /*: {|
     hasTable: boolean,
-    paramSet: {...},
-    runtimeArgs: Array<BabelNodeCallExpression>,
+    paramSet: ParamSet,
+    runtimeArgs: $ReadOnlyArray<BabelNodeCallExpression>,
     fileSource: string,
     usedEnums: {[enumExpr: string]: ?boolean},
-    variations: {...},
+    variations: {[paramName: string]:
+      {|
+          type: 'number',
+          token: string,
+        |}
+      | {|
+          type: 'gender',
+          token: string,
+        |}
+    },
   |} */ {
     const {
       fileSource,
@@ -315,9 +338,10 @@ class FbtFunctionCallProcessor {
     switch (textNode.type) {
       case 'ArrayExpression':
         return textNode;
-      case 'BinaryExpression':
+      case 'BinaryExpression': {
         const operands = this._getBinaryExpressionOperands(textNode);
         return arrayExpression(operands);
+      }
       case 'CallExpression':
       case 'StringLiteral':
       case 'TemplateLiteral':
@@ -336,7 +360,7 @@ class FbtFunctionCallProcessor {
     switch (node.type) {
       case 'BinaryExpression':
         if (node.operator !== '+') {
-          throw new Error('Expected to see a string concatenation');
+          throw errorAt(node, 'Expected to see a string concatenation');
         }
         return [
           ...this._getBinaryExpressionOperands(node.left),
@@ -461,7 +485,7 @@ class FbtFunctionCallProcessor {
             value: getRawSource(fileSource, arg0),
           });
           break;
-        case 'plural':
+        case 'plural': {
           if (arg0.type !== 'StringLiteral') {
             throw errorAt(
               arg0,
@@ -521,7 +545,8 @@ class FbtFunctionCallProcessor {
           // $FlowFixMe An object literal isn't compatible with BabelNodeExpression
           texts.push(data);
           break;
-        case 'pronoun':
+        }
+        case 'pronoun': {
           // Usage: fbt.pronoun(usage, gender [, options])
           const optionsNode = arg2;
           const options = collectOptions(
@@ -544,6 +569,7 @@ class FbtFunctionCallProcessor {
           // $FlowFixMe An object literal isn't compatible with BabelNodeExpression
           texts.push(pronounData);
           break;
+        }
         case 'name':
           texts.push(variations[
             // $FlowFixMe `value` property is not always present
@@ -620,12 +646,9 @@ class FbtFunctionCallProcessor {
     const encodedOutput = pluginOptions.fbtBase64
       ? Buffer.from(argsOutput).toString('base64')
       : argsOutput;
+    const fbtSentinel = pluginOptions.fbtSentinel || SENTINEL;
     const args = [
-      t.stringLiteral(
-        pluginOptions.fbtSentinel +
-          encodedOutput +
-          pluginOptions.fbtSentinel,
-      ),
+      t.stringLiteral(fbtSentinel + encodedOutput + fbtSentinel),
     ];
 
     if (runtimeArgs.length > 0) {